requirements.txt
numpy
pandas
scikit-learn
tensorflow
argparse
cryptography
pillow
pydub
pycryptodome
requests